% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEBplots-package.R
\docType{package}
\name{DEEBplots-package}
\alias{DEEBplots}
\alias{DEEBplots-package}
\title{DEEBplots: Create Plots for the DEEB Project}
\description{
Create plots for the DEEB project.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
